Build a way that packages can drop scripts of their own in /etc/openapp-backup/scripts.d which will be run by openapp-backup.
Also, if no files exist at all. At least run all the scripts in the modulesdir
of openapp-backup. We're not really sure what we're backing up, but something
is better than nothing.

After openapp-restore ran the etc-module (which is hardcoded in front of the
line, right after dpkg-selections), refresh the list of available modules. dpkg
should have installed all the packages from the old box, and etc should have
fixed the symlinks for openapp-backup.

Create Symlinks in /etc/openapp-backup/scripts.d/ in the postinst of openapp-backup